requirements-wheels.txt
click == 8.1.*
Flask == 2.2.*
imutils == 0.5.*
matplotlib == 3.6.*
mypy == 0.942
numpy == 1.22.*
opencv-python-headless == 4.5.5.*
paho-mqtt == 1.6.*
peewee == 3.15.*
peewee_migrate == 1.6.*
psutil == 5.9.*
pydantic == 1.10.*
PyYAML == 6.0
pytz == 2022.7
tzlocal == 4.2
types-PyYAML == 6.0.*
requests == 2.28.*
types-requests == 2.28.*
scipy == 1.8.*
setproctitle == 1.3.*
ws4py == 0.5.*
zeroconf == 0.47.*
onnxruntime-gpu
tensorrt
cuda-python == 11.7
# pytorch
nvidia-pyindex; platform_machine == 'x86_64'
# nvidia-tensorrt == 8.4.1.3; platform_machine == 'x86_64'
cuda-python == 11.7; platform_machine == 'x86_64'
cython == 0.29.*; platform_machine == 'x86_64'
nvidia-cuda-runtime-cu11 == 11.7.*; platform_machine == 'x86_64'
nvidia-cublas-cu11 == 11.11.*; platform_machine == 'x86_64'
nvidia-cudnn-cu11 == 8.7.*; platform_machine == 'x86_64'
nvidia-cuda-nvrtc-cu11 == 11.7.*; platform_machine == 'x86_64'
# Openvino Library - Custom built with MYRIAD support
# openvino @ https://github.com/NateMeyer/openvino-wheels/releases/download/multi-arch_2022.2.0/openvino-2022.2.0-000-cp39-cp39-manylinux_2_31_x86_64.whl; platform_machine == 'x86_64'
# openvino @ https://github.com/NateMeyer/openvino-wheels/releases/download/multi-arch_2022.2.0/openvino-2022.2.0-000-cp39-cp39-linux_aarch64.whl; platform_machine == 'aarch64'
# openvino @ https://github.com/NateMeyer/openvino-wheels/releases/download/multi-arch_2022.2.0/openvino-2022.2.0-000-cp39-cp39-linux_armv7l.whl; platform_machine == 'armv7l'
